update
- Changes to SanitizeFilename from #1061 got lost in the merges, causing this PR to fail to build (I only tested on Linux)
- Changes to BR_Win32_WritePrivateProfileString from #1037 also got lost in the process (looks like d0191d1 moved the keyName check to another function)
- minor tweaks

https://github.com/reaper-oss/sws/pull/1092#issuecomment-455888576

+++ nofish/nofish.cpp
@@ -297,7 +297,7 @@ void CycleMIDIRecordingModes(COMMAND_T* ct)
 
 	for (int i = 0; i < selTracks.GetSize(); i++) {
 		MediaTrack* track = selTracks.Get()[i];
-		int recMode = GetMediaTrackInfo_Value(track, "I_RECMODE");
+		int recMode = (int)GetMediaTrackInfo_Value(track, "I_RECMODE");
 		int nextRecMode;
 
 		switch (recMode) {
